requirements.txt
streamlit
streamlit-drawable-canvas
numpy
tensorflow-cpu
scikit-image
pandas
scikit-learn
matplotlib
datetime
seaborn